Move models and schema to their own modules
